split core to other modules
